requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-bootstrap4==23.2
django-crispy-forms==1.14.0
django-filter==23.2
django-import-export==3.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
